+++ lib/ooxml_parser/common_parser/common_data/alternate_content/chart/chart/series/series_text.rb
@@ -20,6 +20,4 @@ module OoxmlParser
       self
     end
   end
-
-  Categories = SeriesText
 end